Create YaoHIRExt
