+++ Project.toml
@@ -7,11 +7,12 @@ version = "0.2.1"
 Expronicon = "6b7a57c9-7cc1-4fdf-b7f5-e857abae3636"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 YaoLocations = "66df03fb-d475-48f7-b449-3d9064bf085b"
-# TODO Move to weak deps, after fixing bug: 
-# ArgumentError: Package YaoHIR does not have OpenQASM in its dependencies:
-OpenQASM = "a8821629-a4c0-4df7-9e00-12969ff383a7"
 
 [weakdeps]
+OpenQASM = "a8821629-a4c0-4df7-9e00-12969ff383a7"
+
+[extensions]
+YaoHIRExt = ["OpenQASM"]
 
 [compat]
 Expronicon = "0.10"
@@ -19,12 +20,9 @@ MLStyle = "0.4"
 YaoLocations = "0.1"
 julia = "1.9"
 
-# FIXME are these dependencies not found?
-#[extensions]
-#BlockIROpenQASMExt = ["OpenQASM", "CompilerPluginTools"]
-
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+OpenQASM = "a8821629-a4c0-4df7-9e00-12969ff383a7"
 
 [targets]
-test = ["Test" ]
+test = ["Test", "OpenQASM"]
